vsync and srgb color format

+++ code/handmade_render_group.h
@@ -5,9 +5,9 @@ struct loaded_bitmap {
 	v2 AlignPercentage;
 	r32 WidthOverHeight;
     
-	s16 Width;
-	s16 Height;
-	s16 Pitch;
+	s32 Width;
+	s32 Height;
+	s32 Pitch;
     
     // opengl handle texture
     u32 Handle;
